Moving utility code to wirepopup file
