+++ federal/libs/global-navigation/src/PostRendering/PopupWiring.ts
@@ -1,12 +1,60 @@
-import {
-  closePopup,
-  togglePopup,
-  isPopupOpen,
-  triggerForPopupId,
-  IS_OPEN_CLASS,
-} from "../Utils/Utils";
-
 const MENU_WRAPPER_ID = 'feds-menu-wrapper';
+
+// Open-state class used in compound selectors (`.feds-popup.is-open`,
+// `.feds-menu-wrapper.is-open`). Replaces the HTML popover API so the gnav
+// stays out of the browser top layer.
+export const IS_OPEN_CLASS = 'is-open';
+
+export const triggerForPopupId = (
+  root: ParentNode,
+  id: string,
+): HTMLElement | null => {
+  if (id === '') return null;
+  return root.querySelector<HTMLElement>(`[aria-controls="${CSS.escape(id)}"]`);
+};
+
+const dispatchToggle = (el: HTMLElement, opening: boolean): void => {
+  const init = {
+    newState: opening ? 'open' : 'closed',
+    oldState: opening ? 'closed' : 'open',
+    bubbles: false,
+    cancelable: false,
+  } as const;
+  // Fall back to a shaped Event for runtimes without ToggleEvent (older jsdom).
+  const ToggleEventCtor = (window as unknown as {
+    ToggleEvent?: typeof ToggleEvent
+  }).ToggleEvent;
+  const event = ToggleEventCtor !== undefined
+    ? new ToggleEventCtor('toggle', init)
+    : Object.assign(new Event('toggle', init), {
+      newState: init.newState,
+      oldState: init.oldState,
+    });
+  el.dispatchEvent(event);
+};
+
+export const isPopupOpen = (el: HTMLElement | null | undefined): boolean =>
+  el !== null && el !== undefined && el.classList.contains(IS_OPEN_CLASS);
+
+const openPopup = (el: HTMLElement | null | undefined): void => {
+  if (el === null || el === undefined) return;
+  if (el.classList.contains(IS_OPEN_CLASS)) return;
+  el.classList.add(IS_OPEN_CLASS);
+  dispatchToggle(el, true);
+};
+
+export const closePopup = (el: HTMLElement | null | undefined): void => {
+  if (el === null || el === undefined) return;
+  if (!el.classList.contains(IS_OPEN_CLASS)) return;
+  el.classList.remove(IS_OPEN_CLASS);
+  dispatchToggle(el, false);
+};
+
+export const togglePopup = (el: HTMLElement | null | undefined): void => {
+  if (el === null || el === undefined) return;
+  if (el.classList.contains(IS_OPEN_CLASS)) closePopup(el);
+  else openPopup(el);
+};
 
 // Click-to-toggle + aria reflection for every popup. Mutual exclusion only
 // applies between mega-menu popups; the hamburger does NOT close them
